REFACTOR: Rename grid-areas with a practical way:
Instead of main-sidebar I choose to go with 'sidebar'.
And for ¿aux-sidebar' or 'contextual-sidebar', I prefer to left it
as 'ads'. why? Because in this framepoint, it's a longshot to try
building a behemot from the begining.

There are 3 layouts in my website:  landing-page (header | main | footer)
blog (header | sidebar | main | ads | footer) and wiki/dashbord (sidebar | main | footer)