Uff, lot's of things happened here:
I tested a few different variations of the rules trying to achieve
a 'Holy-grail' layout with full bleed, combining it with the 'breakout'
layout (no containers).

It happens that, after a lot of exploration and reading documentation
it's just not possible to do it with current CSS rules, as a given grid
can't inherit properties from a Grand parent container.

Thankfully enough, what I have made untill now, works pretty well,
specially the 'LAYOUT VARIATIONS' exploiting the characteristics of
'grid-templates' (as I left explained in the CSS document; wich I will
document using the '.wiki-page' body class, when I finish to develop
the components of my site