REFACTOR: Delete full-width class from various elements:
Aparently those elements are by deafult full-width. It's strange, because in the original breakout-grid
in order to behvae this way, that class is absolutely needed.
In my case, I don't mind.. but this will be a technical debt, because is still a glitch in my code.

Will flag this in the backlog anyways. Better debug now that there are not as many features as I intended
to add in the foreseable future.