STYLE Add styles for buttons
SCOPE: I did a lot of changes and didn't commit properly.
Buut, it was a messy process.. basically I forged 6 rules
2 sets of 3 rules to control style and color
and it workedgit add *
However didn't work too well in 2/3 colors.. just branded.
So.. I decided to refactor into 3 classes controlling
visual appearence (filled, tonal, outline) and default
color as tertiary for primary buttons (filled) and
grey tones or B&W for outline buttons.

Maybe in the future I will find a way to create a better
system. But for now, this will have to do the work.